optimized code refactoring it in general
