+++ src/mandate_data/__init__.py
@@ -1,4 +1,4 @@
-from .mandate_data_get import mandate_data_get_router
-from .mandate_data_post import mandate_data_post_router
-from .mandate_data_put import mandate_data_put_router
-from .mandate_data_delete import mandate_data_delete_router
+from .delete import mandate_data_delete_router
+from .get import mandate_data_get_router
+from .post import mandate_data_post_router
+from .put import mandate_data_put_router
