+++ src/meter_data/__init__.py
@@ -1,4 +1,4 @@
-from .meter_data_get import meter_data_get_router
-from .meter_data_post import meter_data_post_router
-from .meter_data_put import meter_data_put_router
-from .meter_data_delete import meter_data_delete_router
+from .delete import meter_data_delete_router
+from .get import meter_data_get_router
+from .post import meter_data_post_router
+from .put import meter_data_put_router
