+++ src/meter_readings/__init__.py
@@ -1,4 +1,4 @@
-from .meter_readings_get import meter_readings_get_router
-from .meter_readings_post import meter_readings_post_router
-from .meter_readings_put import meter_readings_put_router
-from .meter_readings_delete import meter_readings_delete_router
+from .delete import meter_readings_delete_router
+from .get import meter_readings_get_router
+from .post import meter_readings_post_router
+from .put import meter_readings_put_router
