+++ src/models/__init__.py
@@ -1,6 +1,6 @@
 from .brand_enum_model import BrandEnum
 from .energy_type_enum import EnergyTypeEnum
-from .response_model import Response
 from .mandate_data_model import MandateData
 from .meter_data_model import MeterDataRequest, MeterDataResponse
 from .meter_readings_model import MeterReadingsRequest, MeterReadingsResponse
+from .response_model import Response
