+++ src/models/energy_type_enum.py
@@ -1,5 +1,6 @@
 from enum import Enum
 
+
 class EnergyTypeEnum(str, Enum):
     brand1 = "ELECTRICITY"
     brand2 = "GAS"
