+++ src/router.py
@@ -1,4 +1,5 @@
 from fastapi import APIRouter
+
 from .mandate_data import (mandate_data_get_router, mandate_data_post_router,
                            mandate_data_put_router, mandate_data_delete_router)
 from .meter_data import (meter_data_get_router, meter_data_post_router,
