+++ src/utils/logger.py
@@ -3,6 +3,7 @@ import logging
 DEFAULT_LOG_LEVEL = logging.INFO
 DEFAULT_LOG_FORMAT = '%(asctime)s - %(name)s - %(levelname)s - %(message)s'
 
+
 def setup_logger(name: str, level: int = DEFAULT_LOG_LEVEL, fmt: str = DEFAULT_LOG_FORMAT) -> logging.Logger:
     logger = logging.getLogger(name)
     logger.setLevel(level)
